use <component>.jsx + index.jsx files
