+++ src/common/style.jsx
@@ -1,0 +1,1 @@
+export const classesBoxBorder = 'sm:border sm:rounded sm:border-gray-400 sm:bg-white'
